move greeting components to example folder
